+++ server/api/Models/Square.cs
@@ -4,7 +4,7 @@ namespace api.models;
 public class Square
 {
     public int Id { get; set; }
-    public required double X { get; set; }
-    public required double Y { get; set; }
+    public required int X { get; set; }
+    public required int Y { get; set; }
     public required string Color { get; set; }
 }
